Reintroduce mixed precision

+++ src/maxtext/scratch_code/benchmark_gdn_optimization.py
@@ -240,7 +240,7 @@ def run_comparison():
     # --- CONFIGURATION ---
     if backend == 'tpu':
         # REAL BENCHMARK SETTINGS (Heavy)
-        DTYPE = jnp.float32
+        DTYPE = jnp.bf16
         BATCH = 2
         SEQ_LEN = 4096      
         ITERS = 20
